Added POS, refactored, added tests
Renamed the text_preprocessing file into tokenization, and created another text_preprocessing file where we will run all the functions of this milestione
